requirements.txt
# Core dependencies
biopython>=1.75

# Development dependencies (optional, for contributing)
# pytest>=6.0.0
# pytest-cov>=2.10.0
# black>=21.0.0
# flake8>=3.8.0
# mypy>=0.900
